Address PR #128 review comments
- Move split(at:)/valueAndDerivative(at:) from public BernsteinPolynomial
  to internal ClippableBernsteinPolynomial protocol; BezierClippingPolynomial
  now refines ClippableBernsteinPolynomial
- Simplify BP4/BP5 value(at:) to delegate to valueAndDerivative(at:).0
- Create BezierCurve+Clipping.swift: ControlPolygon, BezierClippingCurve protocol,
  convexHullClipInterval, fatLineClip (now BezierClippingCurve-constrained),
  bezierClipping — all moved from BezierCurve.swift/Utils.swift
- NonlinearBezierCurve now refines BezierClippingCurve instead of carrying
  controlPolygon/controlPolygonBounds directly
- Delete old root-finding machinery (newton, findRootBisection,
  findDistinctRootsCallback, findDistinctRoots) from BernsteinPolynomial.swift
- Delete non-unit-interval polynomial tests (testDegree1-5, etc.)
- Delete testCubicMicro_* micro-benchmark functions
- Collapse testAdversarialIntersections into individual test methods
- Remove accuracy: 1e-14 from QuadraticCurveTests assertion
