+++ BezierKit/BezierKitTests/PerformanceTests.swift
@@ -129,119 +129,6 @@ class PerformanceTests: XCTestCase {
         }
     }
 
-    // MARK: - Component microbenchmarks (same 50×50 dataset, reseed 2)
-
-    func testCubicMicro_BoundingBox() {
-        // Measures: tight boundingBox for 2500 pairs (initial bbox overlap check cost).
-        let curves = generateRandomCurves(count: 50, reseed: 2)
-        self.measure {
-            var sink = false
-            for c1 in curves {
-                let b1 = c1.boundingBox
-                for c2 in curves {
-                    sink = sink || b1.overlaps(c2.boundingBox)
-                }
-            }
-            _ = sink
-        }
-    }
-
-    func testCubicMicro_SplitFromTo() {
-        // Measures: split(from:to:) for a mid-range subinterval, 2500 calls.
-        let curves = generateRandomCurves(count: 50, reseed: 2)
-        self.measure {
-            var sink = CGFloat(0)
-            for c1 in curves {
-                for _ in curves {
-                    let s = c1.split(from: 0.25, to: 0.75)
-                    sink += s.p0.x
-                }
-            }
-            _ = sink
-        }
-    }
-
-    func testCubicMicro_SplitAt() {
-        // Measures: split(at:) for comparison with split(from:to:), 2500 calls.
-        let curves = generateRandomCurves(count: 50, reseed: 2)
-        self.measure {
-            var sink = CGFloat(0)
-            for c1 in curves {
-                for _ in curves {
-                    let (left, _) = c1.split(at: 0.5)
-                    sink += left.p0.x
-                }
-            }
-            _ = sink
-        }
-    }
-
-    func testCubicMicro_PointAndDerivative() {
-        // Measures: point(at:) + derivative(at:) twice, matching the work inside split(from:to:).
-        let curves = generateRandomCurves(count: 50, reseed: 2)
-        self.measure {
-            var sink = CGFloat(0)
-            for c1 in curves {
-                for _ in curves {
-                    let p0 = c1.point(at: 0.25)
-                    let p3 = c1.point(at: 0.75)
-                    let d0 = c1.derivative(at: 0.25)
-                    let d1 = c1.derivative(at: 0.75)
-                    sink += p0.x + p3.x + d0.x + d1.x
-                }
-            }
-            _ = sink
-        }
-    }
-
-    func testCubicMicro_ConvexHullClip() {
-        // Measures: Utils.convexHullClipInterval for 4 control points (cubic), 2500 calls.
-        // This is the innermost loop of fatLineClip.
-        let curves = generateRandomCurves(count: 50, reseed: 2)
-        // Build representative d-value arrays (distances from a chord)
-        srand48(99)
-        let samples: [(CGFloat, CGFloat, CGFloat, CGFloat)] = (0..<50).map { _ in
-            (CGFloat(drand48()-0.5), CGFloat(drand48()-0.5),
-             CGFloat(drand48()-0.5), CGFloat(drand48()-0.5))
-        }
-        let dLow: CGFloat = -0.1, dHigh: CGFloat = 0.1
-        self.measure {
-            var hitCount = 0
-            for (i, _) in curves.enumerated() {
-                let (d0, d1, d2, d3) = samples[i]
-                for _ in curves {
-                    if Utils.convexHullClipInterval(d0: d0, d1: d1, d2: d2, d3: d3, n: 4,
-                                                   dLow: dLow, dHigh: dHigh) != nil {
-                        hitCount += 1
-                    }
-                }
-            }
-            _ = hitCount
-        }
-    }
-
-    func testCubicMicro_BezierClippingOnePair() {
-        // Measures: a single bezierClipping call on a pair that DOES intersect.
-        // This isolates the cost of the full clipping convergence for one typical pair.
-        let curves = generateRandomCurves(count: 50, reseed: 2)
-        // Find a pair that actually intersects so the clipping runs to convergence.
-        var c1: CubicCurve = curves[0], c2: CubicCurve = curves[0]
-        outer: for a in curves { for b in curves where a != b {
-            if !a.intersections(with: b, accuracy: 1e-5).isEmpty { c1=a; c2=b; break outer }
-        }}
-        let sub1 = Subcurve(curve: c1), sub2 = Subcurve(curve: c2)
-        self.measure {
-            var total = 0
-            for _ in 0..<2500 {
-                var result: [Intersection] = []
-                var iters = 0
-                _ = Utils.bezierClipping(sub1, sub2, &result, &iters)
-                total += result.count
-            }
-            _ = total
-        }
-    }
-
     func testCubicIntersectionsPerformanceTangentEndpoint() {
         // test the performance of `intersections(with:,accuracy:)`
         // -Onone 0.89 seconds
