+++ BezierKit/BezierKitTests/PolynomialTests.swift
@@ -19,7 +19,7 @@ class PolynomialTests: XCTestCase {
         XCTAssertEqual(point.value(at: 0.5), 3)
         XCTAssertEqual(point.value(at: 1), 3)
         XCTAssertEqual(point.derivative, BernsteinPolynomial0(b0: 0.0))
-        XCTAssertEqual(findDistinctRoots(of: point, between: 0, and: 1), [])
+        XCTAssertEqual(findDistinctRootsInUnitInterval(of: point), [])
         XCTAssertEqual(point.b0, 3.0)
 
         let line = BernsteinPolynomial1(b0: 2.0, b1: 4.0)
@@ -27,8 +27,7 @@ class PolynomialTests: XCTestCase {
         XCTAssertEqual(line.value(at: 0.5), 3)
         XCTAssertEqual(line.value(at: 1), 4)
         XCTAssertEqual(line.derivative, BernsteinPolynomial0(b0: 2))
-        XCTAssertEqual(findDistinctRoots(of: line, between: -2, and: 1), [-1])
-        XCTAssertEqual(findDistinctRoots(of: line, between: 0, and: 1), [])
+        XCTAssertEqual(findDistinctRootsInUnitInterval(of: line), [])
         XCTAssertEqual(line.b0, 2)
         XCTAssertEqual(line.b1, 4)
 
@@ -42,42 +41,6 @@ class PolynomialTests: XCTestCase {
         XCTAssertEqual(quad.b2, 0)
     }
 
-    func testDegree1() {
-        let polynomial = BernsteinPolynomial1(b0: -3, b1: 2)
-        let roots = findDistinctRoots(of: polynomial, between: -1, and: 1)
-        XCTAssertEqual(roots.count, 1)
-        XCTAssertEqual(roots[0], CGFloat(0.6), accuracy: accuracy)
-    }
-
-    func testDegree2() {
-        let polynomial = BernsteinPolynomial2(b0: -5, b1: -6, b2: -4)
-        let roots = findDistinctRoots(of: polynomial, between: -10, and: 10)
-        XCTAssertEqual(roots[0], -1, accuracy: accuracy)
-        XCTAssertEqual(roots[1], 1.0 + 2.0 / 3.0, accuracy: accuracy)
-    }
-
-    func testDegree3() {
-        // x^3 - 6x^2 + 11x - 6
-        let polynomial = BernsteinPolynomial3(b0: -6, b1: -7.0 / 3.0, b2: -2.0 / 3.0, b3: 0)
-        XCTAssertEqual(polynomial.b0, -6)
-        XCTAssertEqual(polynomial.b1, CGFloat(-7.0 / 3.0))
-        XCTAssertEqual(polynomial.b2, CGFloat(-2.0 / 3.0))
-        XCTAssertEqual(polynomial.b3, 0.0)
-        let roots = findDistinctRoots(of: polynomial, between: 0, and: 4)
-        XCTAssertEqual(roots[0], 1, accuracy: accuracy)
-        XCTAssertEqual(roots[1], 2, accuracy: accuracy)
-        XCTAssertEqual(roots[2], 3, accuracy: accuracy)
-    }
-
-    func testDegree3RepeatedRoot1() {
-        // x^3 - 4x^2 + 5x - 2
-        // repeated root at x = 1
-        let polynomial = BernsteinPolynomial3(b0: -2, b1: -1.0 / 3.0, b2: 0, b3: 0)
-        let roots = findDistinctRoots(of: polynomial, between: -1, and: 3)
-        XCTAssertEqual(roots[0], 1, accuracy: accuracy)
-        XCTAssertEqual(roots[1], 2, accuracy: accuracy)
-    }
-
     func testDegree3RootExactlyZero() {
         // root is exactly t = 0 (at the start of unit interval),
         // so may be accidentally discarded due to numerical precision
@@ -86,49 +49,6 @@ class PolynomialTests: XCTestCase {
         XCTAssertEqual(roots.count, 2)
         XCTAssertEqual(roots[0], 0.0)
         XCTAssertEqual(roots[1], 2.0 / 3.0, accuracy: accuracy)
-    }
-
-    func testDegree4() {
-        // x^4 - 2.44x^2 + 1.44
-        let polynomial = BernsteinPolynomial4(b0: 1.44, b1: 1.44, b2: CGFloat(1.44 - 1.22 / 3), b3: 0.22, b4: 0)
-        XCTAssertEqual(polynomial.b0, 1.44)
-        XCTAssertEqual(polynomial.b1, 1.44)
-        XCTAssertEqual(polynomial.b2, CGFloat(1.44 - 1.22 / 3))
-        XCTAssertEqual(polynomial.b3, 0.22)
-        XCTAssertEqual(polynomial.b4, 0)
-        let roots = findDistinctRoots(of: polynomial, between: -2, and: 2)
-        XCTAssertEqual(roots[0], -1.2, accuracy: accuracy)
-        XCTAssertEqual(roots[1], -1, accuracy: accuracy)
-        XCTAssertEqual(roots[2], 1, accuracy: accuracy)
-        XCTAssertEqual(roots[3], 1.2, accuracy: accuracy)
-    }
-    // Skip on platforms where CGFloat is 32bit
-    #if !(arch(i386) || arch(arm) || arch(wasm32))
-    func testDegree4RepeatedRoots() {
-        // x^4 - 2x^2 + 1
-        // derivative 4x^3 - 4x should produce intervals of -2, -1, 0, 1, 2
-        let polynomial = BernsteinPolynomial4(b0: 1, b1: 1, b2: 2.0 / 3.0, b3: 0, b4: 0)
-        let roots = findDistinctRoots(of: polynomial, between: -2, and: 2)
-        XCTAssertEqual(roots.count, 2)
-        XCTAssertEqual(roots[0], -1, accuracy: accuracy)
-        if roots.count > 1 {
-            XCTAssertEqual(roots[1], 1, accuracy: accuracy)
-        }
-    }
-    #endif
-    func testDegree5() {
-        // 0.2x^5 - 0.813333x^3 - 8.56x
-        let polynomial = BernsteinPolynomial5(b0: 0, b1: -1.712, b2: -3.424, b3: -5.2173333, b4: -7.1733332, b5: -9.173333)
-        XCTAssertEqual(polynomial.b0, 0)
-        XCTAssertEqual(polynomial.b1, -1.712)
-        XCTAssertEqual(polynomial.b2, -3.424)
-        XCTAssertEqual(polynomial.b3, -5.2173333)
-        XCTAssertEqual(polynomial.b4, -7.1733332)
-        XCTAssertEqual(polynomial.b5, -9.173333)
-        let roots = findDistinctRoots(of: polynomial, between: -4, and: 4)
-        XCTAssertEqual(roots[0], -2.9806382, accuracy: accuracy)
-        XCTAssertEqual(roots[1], 0, accuracy: accuracy)
-        XCTAssertEqual(roots[2], 2.9806382, accuracy: accuracy)
     }
 
     func testDegree4RealWorldIssue() {
