+++ BezierKit/Library/BernsteinPolynomial.swift
@@ -19,9 +19,31 @@ public protocol BernsteinPolynomial: Equatable, Sendable {
     associatedtype NextLowerOrderPolynomial: BernsteinPolynomial
     func value(at x: CGFloat) -> CGFloat
     var derivative: NextLowerOrderPolynomial { get }
-    /// Returns (value, derivative) at x. Conformers can override to share de Casteljau intermediates.
+}
+
+/// Internal protocol for types that also support splitting and combined value/derivative evaluation.
+/// BezierClippingPolynomial refines this; the concrete BP types all satisfy it.
+internal protocol ClippableBernsteinPolynomial: BernsteinPolynomial {
     func valueAndDerivative(at x: CGFloat) -> (CGFloat, CGFloat)
     func split(at t: CGFloat) -> (left: Self, right: Self)
+}
+
+internal extension ClippableBernsteinPolynomial {
+    func valueAndDerivative(at x: CGFloat) -> (CGFloat, CGFloat) {
+        (value(at: x), derivative.value(at: x))
+    }
+    func split(from t1: CGFloat, to t2: CGFloat) -> Self {
+        guard t1 != 0 else { return split(at: t2).left }
+        let right = split(at: t1).right
+        guard t2 != 1 else { return right }
+        return right.split(at: (t2 - t1) / (1 - t1)).left
+    }
+}
+
+public extension BernsteinPolynomial {
+    func valueAndDerivative(at x: CGFloat) -> (CGFloat, CGFloat) {
+        (value(at: x), derivative.value(at: x))
+    }
 }
 
 extension BernsteinPolynomial0: AnalyticalRootsCallback {
@@ -52,18 +74,6 @@ extension BernsteinPolynomial3: AnalyticalRootsCallback {
             guard $0 >= start, $0 <= end else { return }
             callback($0)
         }
-    }
-}
-
-public extension BernsteinPolynomial {
-    func valueAndDerivative(at x: CGFloat) -> (CGFloat, CGFloat) {
-        (value(at: x), derivative.value(at: x))
-    }
-    func split(from t1: CGFloat, to t2: CGFloat) -> Self {
-        guard t1 != 0 else { return split(at: t2).left }
-        let right = split(at: t1).right
-        guard t2 != 1 else { return right }
-        return right.split(at: (t2 - t1) / (1 - t1)).left
     }
 }
 
@@ -162,12 +172,7 @@ public struct BernsteinPolynomial4: BernsteinPolynomial {
         BernsteinPolynomial3(b0: 4 * (b1 - b0), b1: 4 * (b2 - b1), b2: 4 * (b3 - b2), b3: 4 * (b4 - b3))
     }
     public func value(at x: CGFloat) -> CGFloat {
-        let s = 1 - x, t = x
-        let c10 = s * b0 + t * b1; let c11 = s * b1 + t * b2
-        let c12 = s * b2 + t * b3; let c13 = s * b3 + t * b4
-        let c20 = s * c10 + t * c11; let c21 = s * c11 + t * c12; let c22 = s * c12 + t * c13
-        let c30 = s * c20 + t * c21; let c31 = s * c21 + t * c22
-        return s * c30 + t * c31
+        valueAndDerivative(at: x).0
     }
     public func valueAndDerivative(at x: CGFloat) -> (CGFloat, CGFloat) {
         let s = 1 - x, t = x
@@ -207,6 +212,19 @@ public struct BernsteinPolynomial5: BernsteinPolynomial {
     public var derivative: BernsteinPolynomial4 {
         BernsteinPolynomial4(b0: 5 * (b1 - b0), b1: 5 * (b2 - b1), b2: 5 * (b3 - b2), b3: 5 * (b4 - b3), b4: 5 * (b5 - b4))
     }
+    public func value(at x: CGFloat) -> CGFloat {
+        valueAndDerivative(at: x).0
+    }
+    public func valueAndDerivative(at x: CGFloat) -> (CGFloat, CGFloat) {
+        let s = 1 - x, t = x
+        let c10 = s * b0 + t * b1; let c11 = s * b1 + t * b2
+        let c12 = s * b2 + t * b3; let c13 = s * b3 + t * b4; let c14 = s * b4 + t * b5
+        let c20 = s * c10 + t * c11; let c21 = s * c11 + t * c12
+        let c22 = s * c12 + t * c13; let c23 = s * c13 + t * c14
+        let c30 = s * c20 + t * c21; let c31 = s * c21 + t * c22; let c32 = s * c22 + t * c23
+        let c40 = s * c30 + t * c31; let c41 = s * c31 + t * c32
+        return (s * c40 + t * c41, 5 * (c41 - c40))
+    }
     public func split(at t: CGFloat) -> (left: BernsteinPolynomial5, right: BernsteinPolynomial5) {
         let h00 = Utils.linearInterpolate(b0, b1, t)
         let h01 = Utils.linearInterpolate(b1, b2, t)
@@ -226,135 +244,6 @@ public struct BernsteinPolynomial5: BernsteinPolynomial {
         return (left: BernsteinPolynomial5(b0: b0, b1: h00, b2: h10, b3: h20, b4: h30, b5: h40),
                 right: BernsteinPolynomial5(b0: h40, b1: h31, b2: h22, b3: h13, b4: h04, b5: b5))
     }
-    public func value(at x: CGFloat) -> CGFloat {
-        let s = 1 - x, t = x
-        let c10 = s * b0 + t * b1; let c11 = s * b1 + t * b2
-        let c12 = s * b2 + t * b3; let c13 = s * b3 + t * b4; let c14 = s * b4 + t * b5
-        let c20 = s * c10 + t * c11; let c21 = s * c11 + t * c12
-        let c22 = s * c12 + t * c13; let c23 = s * c13 + t * c14
-        let c30 = s * c20 + t * c21; let c31 = s * c21 + t * c22; let c32 = s * c22 + t * c23
-        let c40 = s * c30 + t * c31; let c41 = s * c31 + t * c32
-        return s * c40 + t * c41
-    }
-    public func valueAndDerivative(at x: CGFloat) -> (CGFloat, CGFloat) {
-        let s = 1 - x, t = x
-        let c10 = s * b0 + t * b1; let c11 = s * b1 + t * b2
-        let c12 = s * b2 + t * b3; let c13 = s * b3 + t * b4; let c14 = s * b4 + t * b5
-        let c20 = s * c10 + t * c11; let c21 = s * c11 + t * c12
-        let c22 = s * c12 + t * c13; let c23 = s * c13 + t * c14
-        let c30 = s * c20 + t * c21; let c31 = s * c21 + t * c22; let c32 = s * c22 + t * c23
-        let c40 = s * c30 + t * c31; let c41 = s * c31 + t * c32
-        return (s * c40 + t * c41, 5 * (c41 - c40))
-    }
-}
-
-private func newton<P: BernsteinPolynomial>(polynomial: P, derivative: P.NextLowerOrderPolynomial, guess: CGFloat, relaxation: CGFloat = 1) -> CGFloat {
-    let maxIterations = 20
-    var x = guess
-    for _ in 0..<maxIterations {
-        let (f, fPrime) = polynomial.valueAndDerivative(at: x)
-        guard f != 0.0 else { break }
-        let delta = relaxation * f / fPrime
-        let previous = x
-        x -= delta
-        guard Swift.abs(x - previous) > 1.0e-10 else { break }
-    }
-    return x
-}
-
-private func findRootBisection<P: BernsteinPolynomial>(of polynomial: P, start: CGFloat, end: CGFloat) -> CGFloat {
-    var guess = (start + end) / 2
-    var low = start
-    var high = end
-    let lowSign = polynomial.value(at: low).sign
-    #if DEBUG
-    let highSign = polynomial.value(at: high).sign
-    assert(lowSign != highSign)
-    #endif
-    let maxIterations = 20
-    var iterations = 0
-    while high - low > 1.0e-5 {
-        let midGuess = (low + high) / 2
-        guess = midGuess
-        let nextGuessF = polynomial.value(at: guess)
-        if nextGuessF == 0 {
-            return guess
-        } else if nextGuessF.sign == lowSign {
-            low = guess
-        } else {
-            #if DEBUG
-            assert(nextGuessF.sign == highSign)
-            #endif
-            high = guess
-        }
-        iterations += 1
-        guard iterations < maxIterations else { break }
-    }
-    return guess
-}
-
-// Zero-allocation root finder for arbitrary intervals. Calls `callback` for each root in sorted order.
-// For degree ≤ 3, delegates to Utils.droots (analytical). For higher degrees,
-// finds critical points of the derivative recursively and searches each interval.
-// Intervals are stored in a withUnsafeTemporaryAllocation buffer (stack for small sizes).
-// Capacity is 6: BernsteinPolynomial5 is the highest-order type defined in this library,
-// and a degree-5 polynomial needs at most [start] + 4 critical points + [end] = 6 slots.
-// A compile-time constant lets the compiler stack-allocate the buffer without a dynamic alloca.
-func findDistinctRootsCallback<P: BernsteinPolynomial>(
-    of polynomial: P,
-    between start: CGFloat, and end: CGFloat,
-    _ callback: (CGFloat) -> Void
-) {
-    if let analytical = polynomial as? AnalyticalRootsCallback {
-        analytical.forEachAnalyticalDistinctRoot(between: start, and: end, callback)
-        return
-    }
-    let derivative = polynomial.derivative
-    withUnsafeTemporaryAllocation(of: CGFloat.self, capacity: 6) { buffer in
-        var count = 0
-        buffer[count] = start; count += 1
-        findDistinctRootsCallback(of: derivative, between: start, and: end) { criticalPoint in
-            buffer[count] = criticalPoint; count += 1
-        }
-        buffer[count] = end; count += 1
-        var lastFoundRoot: CGFloat?
-        for i in 0..<count - 1 {
-            let iStart = buffer[i]
-            let iEnd   = buffer[i + 1]
-            let fStart = polynomial.value(at: iStart)
-            let fEnd   = polynomial.value(at: iEnd)
-            let absFStart = Swift.abs(fStart)
-            let absFEnd = Swift.abs(fEnd)
-            let scale = Swift.max(absFStart, absFEnd)
-            let residualToConsiderRoot = scale * CGFloat.ulpOfOne.squareRoot()
-            let root: CGFloat
-            if fStart * fEnd < 0 {
-                let guess = (iStart + iEnd) / 2
-                let newtonRoot = newton(polynomial: polynomial, derivative: derivative, guess: guess)
-                if iStart < newtonRoot, newtonRoot < iEnd,
-                   Swift.abs(polynomial.value(at: newtonRoot)) <= residualToConsiderRoot {
-                    root = newtonRoot
-                } else {
-                    root = findRootBisection(of: polynomial, start: iStart, end: iEnd)
-                }
-            } else if absFStart <= residualToConsiderRoot, absFEnd >= residualToConsiderRoot {
-                root = iStart
-            } else if absFStart > residualToConsiderRoot, absFEnd <= residualToConsiderRoot {
-                root = iEnd
-            } else {
-                continue
-            }
-            if let lastFoundRoot, lastFoundRoot + 1.0e-5 >= root { continue }
-            lastFoundRoot = root
-            callback(root)
-        }
-    }
-}
-
-func findDistinctRoots<P: BernsteinPolynomial>(of polynomial: P, between start: CGFloat, and end: CGFloat) -> [CGFloat] {
-    var result: [CGFloat] = []
-    findDistinctRootsCallback(of: polynomial, between: start, and: end) { result.append($0) }
-    return result
 }
 
 // Finds roots in [0, 1]. Uses analytical formulas for degree ≤ 3, bezier clipping for degree 4–5.
